rename package idlenotifier to onidle
