+++ pkg/networkservice/common/onidle/option.go
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package idlenotifier
+package onidle
 
 import (
 	"time"
